+++ HISTORY.rst
@@ -6,6 +6,7 @@ unreleased
 
 * *new:* set log format per :envvar:LOG_FORMAT
 * *new:* an options label to set behavioural flags
+* *new:* containers can be identified as a service by configurable labels
 
 0.1.beta1 (2016-12-04)
 ----------------------